requirements.txt
torch
torchvision
transformers
datasets
pillow
pandas
pyyaml
tqdm
scikit-learn
open_clip_torch
numpy
pyarrow
faiss-cpu
